Enhance book search with Goodreads description
