+++ src/books/openLibraryInvoke.ts
@@ -1,13 +1,12 @@
-import { EmbedBuilder, Message } from "discord.js";
+import { Message } from "discord.js";
 import { isSendableChannel } from "../util/typeGuards";
 import {
   BOOKS_SEARCH_OPENLIBRARY_URL,
   REGEX_BOOKS_PATTERN,
-  REGEX_DISCORD_MESSAGE_LENGTH_SHORT,
 } from "../consts/constants";
 import { OpenLibraryTypes } from "../types/books/OpenLibraryResponse";
-import { wrapCodeBlockString } from "../util/commonFunctions";
 import { openLibraryShowBookList } from "./openLibraryShowBookList";
+import { openBooksFound } from "./openLibraryBookFound";
 
 export async function openLibraryInvoke(message: Message): Promise<void> {
   if (!isSendableChannel(message.channel)) return;
@@ -40,6 +39,10 @@ export async function openLibrarySearch(
 ): Promise<void> {
   if (!isSendableChannel(message.channel)) return;
 
+  const replyMessage = await message.reply(
+    `Fetching \`${input}, ${author ?? "any author"}\`...`
+  );
+
   let url = `${BOOKS_SEARCH_OPENLIBRARY_URL}${encodeURIComponent(input)}`;
   if (author) {
     url += `&author:${encodeURIComponent(author)}`;
@@ -50,7 +53,7 @@ export async function openLibrarySearch(
     .then((json: OpenLibraryTypes.Response) => json.docs);
 
   if (!results || results.length === 0) {
-    message.channel.send(
+    replyMessage.edit(
       `Could not find book from query \`${input}\` by \`${
         author ?? "any author"
       }\``
@@ -59,68 +62,9 @@ export async function openLibrarySearch(
   }
 
   if (results.length > 1 && searchMultiple) {
-    await openLibraryShowBookList(message, results, input);
+    await openLibraryShowBookList(message, replyMessage, results, input);
     return;
   }
 
-  await openBooksFound(message, results[0]);
-}
-
-export async function openBooksFound(
-  message: Message,
-  book: OpenLibraryTypes.Book
-): Promise<void> {
-  if (!isSendableChannel(message.channel)) return;
-
-  if (message.content.slice(-1) === "*") {
-    const debugOutput = JSON.stringify(book, null, 4);
-    const outputArray: RegExpMatchArray | [] =
-      debugOutput.match(REGEX_DISCORD_MESSAGE_LENGTH_SHORT) ?? [];
-    for (const r of outputArray) {
-      await message.channel.send(wrapCodeBlockString(r, "json"));
-    }
-    return;
-  }
-
-  const workInfo: OpenLibraryTypes.Work = await fetch(
-    `https://openlibrary.org${book.key}.json`
-  ).then((response: Response) => response.json());
-
-  const authorString = book.author_name?.join(", ") ?? "Unknown Authors";
-
-  const embed: EmbedBuilder = new EmbedBuilder()
-    .setTitle(book.title)
-    .setURL(`https://openlibrary.org${book.key}`)
-    .setDescription(authorString)
-    .setImage(
-      `https://covers.openlibrary.org/b/olid/${book.cover_edition_key}-L.jpg`
-    )
-    .addFields(
-      [
-        workInfo.description
-          ? {
-              name: "Description",
-              value: getDescription(workInfo.description),
-            }
-          : null,
-        workInfo.subjects
-          ? {
-              name: "Tags",
-              value: workInfo.subjects.slice(0, 5).join(", "),
-            }
-          : null,
-      ].filter((a) => a !== null)
-    );
-
-  await message.channel.send({
-    embeds: [embed],
-  });
-}
-
-function getDescription(description: OpenLibraryTypes.Work["description"]) {
-  const descriptionString =
-    typeof description === "string" ? description : description!.value;
-  return descriptionString.length > 320
-    ? descriptionString.slice(0, 320) + "..."
-    : descriptionString;
+  await openBooksFound(message, replyMessage, results[0]);
 }
